Refactor leave request management: remove description from department payload, implement half-day leave functionality with session selection, and enhance leave request validation. Update related models and controllers to support new leave duration options and improve overall user experience in leave management.

+++ server/models/dashboardModel.js
@@ -141,7 +141,7 @@ export async function getEmployeeDashboardStats(employeeId) {
           AND status IN ('Present', 'Half Day')
       ) AS "avgHoursMonth",
       (
-        SELECT COALESCE(SUM(leave_days), 0)::int
+        SELECT COALESCE(SUM(leave_days), 0)::float
         FROM leave_requests
         WHERE employee_id = $1
           AND status = 'Approved'
